improved bullet firing (should add weapon animation though). Added bullet collision with wall and player. Added random respawner and respawn points in map. Should add it to the gameflow

+++ VirtualCommando/src/main/java/me/martin/radev/game/virtualcommando/game/weapon/bullet/NormalBullet.java
@@ -16,7 +16,7 @@ import me.martin.radev.game.virtualcommando.view.graphics.entity.GraphicalRectan
  */
 public class NormalBullet extends Bullet {
     
-    private static final double DAMAGE = 5d;
+    private static final int DAMAGE = 5;
     private static final double VELOCITY = 2.5d;
     private static final double BULLET_WIDTH = 5d;
     private static final double BULLET_HEIGHT = 5d;
